+++ mmc/trunk/src/mcx_utils.h
@@ -26,6 +26,7 @@
 #define _MMC_UTILITIES_H
 
 #include <stdio.h>
+#include <stdint.h>
 #include <vector_types.h>
 #include "cjson/cJSON.h"
 
@@ -173,7 +174,7 @@ typedef struct MMC_config{
     void *photonseed;
     float *replayweight;
     float *replaytime;
-    float *replaydetidx;	/** for replay under wide-field detection mode, record the idx of the pixel where the photon arrives*/
+    uint32_t *replaydetidx;	/** for replay under wide-field detection mode, record the idx of the pixel where the photon arrives*/
     char seedfile[MAX_PATH_LENGTH];
 } mcconfig;
